2019-10-21_Fix EffectData.id type error

+++ Assets/Editor/Script/CSharp/SkillEditor/Model/LuaAnimClipStructure.cs
@@ -404,7 +404,7 @@ namespace SkillEditor.LuaStructure {
                     type = (short)(int)value;
                     return;
                 case Key_Id:
-                    id = (short)(int)value;
+                    id = (int)value;
                     return;
             }
         }
